Allow editing and restoring deleted posts

+++ resources/views/admin/posts/edit.blade.php
@@ -56,9 +56,15 @@
 
     {!! Form::close() !!}
 
-    {!! Form::open(['route' => ['admin.post.destroy', $post], 'method' => 'delete']) !!}
-        <p><button type="submit" class="btn btn-danger">Delete post</button></p>
-    {!! Form::close() !!}
+    @if(!$post->deleted_at)
+        {!! Form::open(['route' => ['admin.post.destroy', $post], 'method' => 'delete']) !!}
+            <p><button type="submit" class="btn btn-danger">Delete post</button></p>
+        {!! Form::close() !!}
+    @else   
+        {!! Form::open(['route' => ['admin.post.restore', $post], 'method' => 'put']) !!}
+            <p><button type="submit" class="btn btn-secondary">Restore post</button></p>
+        {!! Form::close() !!}
+    @endif
 
 @endsection
